<Edit Logo Scene> input the LogoView.cs code in LogoAnimation.cs. Edit LogoView.cs to move next scene. Create SceneController.cs to controll game scene.
