+++ Assets/Views/LogoView/LogoAnimation.cs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b182c133ab3638b499485e6abb781d1a
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
